changes boku

+++ src/user/farmer/farmers.service.ts
@@ -5,17 +5,15 @@ import {
 } from '@nestjs/common';
 import { UserRepository } from '../user.repository';
 import { GetFarmersByTypeDto } from './dtos/get-farmers-by-type.dto';
-import { ErrorService } from 'src/error/error.service';
 import { User } from '@prisma/client';
 import { GetFarmersFromSearchDto } from './dtos/get-farmers-search';
 import { UpdatePasswordDto } from './dtos/update-password.dto';
 import { hash } from 'bcryptjs';
-import { EmailService } from 'src/utils/email.service';
+import { EmailService } from '../../utils/email.service';
 import { VerifyPasswordDto } from './dtos/verify-password-otp.dto';
 
 @Injectable()
 export class FarmersService {
-  private loggerName = 'FarmersService';
   private otpLifeSpan = 300000; // 5 minutes
   private genRandomOtp = (): string => {
     return Math.floor(1000 + Math.random() * 9000).toString();
@@ -23,7 +21,6 @@ export class FarmersService {
 
   constructor(
     private readonly userRepository: UserRepository,
-    private readonly errorService: ErrorService,
     private readonly emailService: EmailService,
   ) {}
 
@@ -31,11 +28,9 @@ export class FarmersService {
     const { farmerType } = getFarmersByTypeDto;
     const where = { farmerType };
     let farmers: User[];
-    try {
-      farmers = await this.userRepository.findUsers({ where });
-    } catch (error) {
-      this.errorService.throwUnexpectedError(error, this.loggerName);
-    }
+
+    farmers = await this.userRepository.findUsers({ where });
+
     farmers = farmers.filter(
       (farmer) => farmer.id !== user.id && farmer.isVerified,
     );
@@ -60,11 +55,8 @@ export class FarmersService {
     };
     let farmers: User[];
 
-    try {
-      farmers = await this.userRepository.findUsers({ where });
-    } catch (error) {
-      this.errorService.throwUnexpectedError(error, this.loggerName);
-    }
+    farmers = await this.userRepository.findUsers({ where });
+
     farmers = farmers.filter(
       (farmer) => farmer.id !== user.id && farmer.isVerified,
     );
@@ -73,32 +65,33 @@ export class FarmersService {
   }
 
   async deleteFarmer(user: User) {
-    let farmer: User;
-    try {
-      farmer = await this.userRepository.deleteUser({ where: { id: user.id } });
-    } catch (error) {
-      if (error.code === 'P2025')
-        throw new NotFoundException(
-          `A user with the id ${user.id} does not exist`,
-        );
-      this.errorService.throwUnexpectedError(error, this.loggerName);
-    }
+    const farmer = await this.userRepository.findOne({
+      where: { id: user.id },
+    });
+
+    if (!farmer)
+      throw new NotFoundException(
+        `A user with the id ${user.id} does not exist`,
+      );
+
+    await this.userRepository.deleteUser({ where: { id: user.id } });
+
     return this.exclude(farmer);
   }
 
   async requestUpdatePassword(email: string) {
+    let farmer = await this.userRepository.findOne({
+      where: { email },
+    });
+
+    if (!farmer) throw new NotFoundException('No user with this email.');
+
     const otp = this.genRandomOtp();
-    let farmer: User;
-    try {
-      farmer = await this.userRepository.editUserInfo({
-        where: { email },
-        data: { otp },
-      });
-    } catch (error) {
-      if (error.code === 'P2025')
-        throw new NotFoundException('No user with this email.');
-      this.errorService.throwUnexpectedError(error, 'FarmerService');
-    }
+    farmer = await this.userRepository.editUserInfo({
+      where: { email },
+      data: { otp },
+    });
+
     await this.emailService.sendOtpForPasswordReset(
       email,
       otp,
